requirements.txt
torch>=2.2.0
torchaudio>=2.2.0
torchcodec>=0.10.0
transformers>=4.41.0
encodec>=0.1.1
mlflow>=2.13.0
fastapi>=0.110.0
uvicorn>=0.29.0
pyyaml>=6.0.1
soundfile>=0.12.1
numpy>=1.26.0
tqdm>=4.66.0
peft>=0.11.1
jiwer>=3.0.4
